add UI for endpoints
user interface to test endpoints

+++ backend/macro-for-geeks-api/Repositories/IUserRepo.cs
@@ -6,7 +6,7 @@ namespace macro_for_geeks_api.Repositories
     public interface IUserRepo
     {
         List<UserViewModel> GetUsers();
-        List<UserViewModel> GetUserById(long id);
+        List<UserViewModel> GetUserById(short id);
       
     }
 }
